GameOfLife: Switch handling of cells
This patch changes the way cells are painted. Previously, BoardWidget
would create an individual Widget for each cell, and recreate the array
when the board would change size. Now there is no more concept of a
"cell widget", the BoardWidget paints itself based on the info it gets
from the Board class.

This patch also switches tracking of cells in the Board class from a manually
managed array of bools to a Vector<bool>.